// src/common/enums/entity.enum.ts
export enum EntityName{
    User = "user",
    Profile = "profile",
    Otp = "otp",
    Category = "category",
    Blog ="blog",
    BlogCategory = "blog_category",
    BlogLike = "blog_like",
    BlogComment = "blog_comment",
    BlogBookmark = "blog_bookmark",
    Image = "image",
    Follow = "follow",
}